+++ data/template53.go
@@ -152,7 +152,7 @@ func (t *Template53) BitsPerValue() uint8 {
 // If bitmap is provided, it must have length equal to the number of grid points.
 // The output will have the same length as the bitmap, with undefined values
 // set to 9.999e20 where bitmap is false.
-func (t *Template53) Decode(packedData []byte, bitmap []bool) ([]float64, error) {
+func (t *Template53) Decode(packedData []byte, bitmap []bool) ([]float32, error) {
 	if len(packedData) == 0 {
 		return nil, fmt.Errorf("no packed data to decode")
 	}
@@ -389,8 +389,8 @@ func (t *Template53) reverseSpatialDifferencing2(diffVals []int32, firstVals []i
 }
 
 // applyScalingWithoutBitmap applies scaling when all values are valid.
-func (t *Template53) applyScalingWithoutBitmap(packedValues []int32) []float64 {
-	values := make([]float64, len(packedValues))
+func (t *Template53) applyScalingWithoutBitmap(packedValues []int32) []float32 {
+	values := make([]float32, len(packedValues))
 	for i, packed := range packedValues {
 		values[i] = t.applyScaling(packed)
 	}
@@ -398,13 +398,13 @@ func (t *Template53) applyScalingWithoutBitmap(packedValues []int32) []float64 {
 }
 
 // applyScalingWithBitmap applies scaling and bitmap.
-func (t *Template53) applyScalingWithBitmap(packedValues []int32, bitmap []bool) ([]float64, error) {
+func (t *Template53) applyScalingWithBitmap(packedValues []int32, bitmap []bool) ([]float32, error) {
 	if len(packedValues) > len(bitmap) {
 		return nil, fmt.Errorf("more packed values (%d) than bitmap entries (%d)",
 			len(packedValues), len(bitmap))
 	}
 
-	values := make([]float64, len(bitmap))
+	values := make([]float32, len(bitmap))
 	packedIdx := 0
 
 	for i := range bitmap {
@@ -431,19 +431,19 @@ func (t *Template53) applyScalingWithBitmap(packedValues []int32, bitmap []bool)
 //
 // Formula: value = (R + X * 2^E) / 10^D
 // This is the same formula as Template 5.0, applied after spatial differencing reversal.
-func (t *Template53) applyScaling(packedValue int32) float64 {
+func (t *Template53) applyScaling(packedValue int32) float32 {
 	// Start with reference value
-	value := float64(t.ReferenceValue)
+	value := float32(t.ReferenceValue)
 
 	// Add scaled packed value: X * 2^E
 	if packedValue != 0 {
-		binaryScale := math.Pow(2.0, float64(t.BinaryScaleFactor))
-		value += float64(packedValue) * binaryScale
+		binaryScale := float32(math.Pow(2.0, float64(t.BinaryScaleFactor)))
+		value += float32(packedValue) * binaryScale
 	}
 
 	// Apply decimal scaling: / 10^D
 	if t.DecimalScaleFactor != 0 {
-		decimalScale := math.Pow(10.0, float64(t.DecimalScaleFactor))
+		decimalScale := float32(math.Pow(10.0, float64(t.DecimalScaleFactor)))
 		value /= decimalScale
 	}
